change in route logout and user

+++ routes/api.php
@@ -8,20 +8,20 @@ use Illuminate\Support\Facades\Route;
 
 Route::namespace('\App\Http\Controllers\API')->group(function (){
     Route::post('/login','AuthController@login');
-    Route::post('/logout','AuthController@logout');
-    Route::get('/user','AuthController@user');
+    Route::post('/logout','AuthController@logout')->middleware('auth:api');
+    Route::get('/user','AuthController@user')->middleware('auth:api');
 });
 Route::get('/test',function (Request $request){
     dd('salam reza');
-});
+})->middleware('auth:api');
 
 Route::group(['prefix' => 'users'],function (){
     Route::namespace('\App\Http\Controllers\API')->group(function (){
-        Route::get('/','UserController@index');
-        Route::get('/{id}','UserController@show');
+        Route::get('/','UserController@index')->middleware('auth:api');
+        Route::get('/{id}','UserController@show')->middleware('auth:api');
         Route::post('/','UserController@store');
-        Route::put('/{id}','UserController@update');
-        Route::delete('/{id}','UserController@destroy');
+        Route::put('/{id}','UserController@update')->middleware('auth:api');
+        Route::delete('/{id}','UserController@destroy')->middleware('auth:api');
     });
 });
